# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Có mặt trên tất cả nền tảng kỹ thuật số hàng đầu như Google Play, iOS, website. SandexCare là ứng dụng dịch vụ hỗ trợ khách hàng giải tỏa những khó khăn trong cuộc sống qua cuộc gọi 1-1 với chuyên viên lắng nghe của SandexCare, đem lại cảm hứng và động lực trong cuộc sống - công ty hàng đầu về sức khỏe tâm lý cộng đồng tại Việt Nam.\nSandexCare sẽ luôn bên bạn mọi lúc khi bạn cần nhất và hãy đến với chúng tôi để cùng nhau lan toả năng lượng tích cực cho cuộc sống để những vấn đề của chính bạn trở nên đơn giản hơn.\n\nFormat date: \"2022-01-04\".\nFormat time_slot: \"00:00-06:00\".\nFormat booking_time: \"01:00\".",
        "title": "Tài liệu đặc tả api.sandexcare.com",
        "termsOfService": "http://sandexcare.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/listener/change-password": {
            "post": {
                "description": "Xử lý thay đổi mật khẩu!",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Listener"
                ],
                "summary": "API change password",
                "parameters": [
                    {
                        "description": "Input Change Password",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputChangePassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/listener/forgot-password": {
            "post": {
                "description": "Xử lý quên mật khẩu!",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Listener"
                ],
                "summary": "API forgot password",
                "parameters": [
                    {
                        "description": "Input Forgot Password",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputForgotPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/listener/resend-forgot-password": {
            "post": {
                "description": "Xử lý gửi lại quên mật khẩu!",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Listener"
                ],
                "summary": "API resend forgot password",
                "parameters": [
                    {
                        "description": "Input Forgot Password",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputForgotPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
        "/listener/reset-new-password": {
            "post": {
                "description": "Xử lý đặt lại mật khẩu mới!",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Listener"
                ],
                "summary": "API reset new password",
                "parameters": [
                    {
                        "description": "Input Reset New Password",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputResetNewPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/listener/verify-reset-password": {
            "post": {
                "description": "Xử lý xác nhận đặt lại mật khẩu!",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Listener"
                ],
                "summary": "API verify reset password",
                "parameters": [
                    {
                        "description": "Input Verify Reset Password",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputVerifyResetPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.InputChangePassword": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "model.InputForgotPassword": {
            "type": "object",
            "properties": {
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
        "model.InputResetNewPassword": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "reset_code": {
                    "type": "string"
                }
            }
        },
...
        "model.InputVerifyResetPassword": {
            "type": "object",
            "properties": {
                "phone_number": {
                    "type": "string"
                },
                "reset_code": {
                    "type": "string"
                }
            }
        },
...
        "model.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.ResponseSuccess": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}